ConDec-359: Integrate methods of PersistenceProvider into AbstractPersistenceManager class in order to reduce the amount of classes
