+++ src/main/java/de/uhd/ifi/se/decision/management/jira/view/treeviewer/TreeViewer.java
@@ -25,7 +25,6 @@ import de.uhd.ifi.se.decision.management.jira.model.KnowledgeType;
 import de.uhd.ifi.se.decision.management.jira.model.Link;
 import de.uhd.ifi.se.decision.management.jira.persistence.AbstractPersistenceManager;
 import de.uhd.ifi.se.decision.management.jira.persistence.GenericLinkManager;
-import de.uhd.ifi.se.decision.management.jira.persistence.PersistenceProvider;
 
 /**
  * Creates tree viewer content
@@ -63,7 +62,7 @@ public class TreeViewer {
 	public TreeViewer(String projectKey, KnowledgeType rootElementType) {
 		this();
 		if (rootElementType != KnowledgeType.OTHER) {
-			AbstractPersistenceManager strategy = PersistenceProvider.getPersistenceStrategy(projectKey);
+			AbstractPersistenceManager strategy = AbstractPersistenceManager.getPersistenceStrategy(projectKey);
 			List<DecisionKnowledgeElement> elements = strategy.getDecisionKnowledgeElements(rootElementType);
 
 			Set<Data> dataSet = new HashSet<Data>();
